+++ changelog.md
@@ -21,6 +21,7 @@
 - Added new compatibility check for Oxygen
 - Added info-page for Pro plugin with option to install the Pro-plugin with valid license key
 - Added new handling for admin notices for better overview over messages from the plugin
+- Added links to edit position settings in Personio if login URL is given
 
 ### Changed